Complete Random World Generation
Generate random coordinates to create rooms without intersection and
then choosing a random coordinate within each room as a gate access
point. Halls between the gates are generated using Kruskal's algorithm,
using the coordinates as nodes and the length of the hall as the weight.
This ensures a reachable world as well as a fun to play one.